+++ .gitignore
@@ -1,4 +1,3 @@
 dist
-types
 node_modules
 results
