+++ experimental/src/cumulative-input-delay.js
@@ -1,7 +1,6 @@
-import { onVisibilityChange } from '../utils/visibility-change'
-import { round } from '../utils/index'
-import { observeEntries } from '../performance-observer'
-import { now } from '../user-timing'
+import { observeEntries, now } from '../..'
+import { onVisibilityChange } from '../../src/utils/visibility-change'
+import { round } from '../../src/utils/index'
 
 /** @typedef {{ metricType: 'cid', value: number, detail: { totalEntries: number, sessionDuration: number } }} CidMetric */
